Join all unit tests in one file #18
All the tests of the 'vfsacrsnh_lib.c' should be in only one file
'Test_vfsacrsnh_lib.c'
